osutil/disks: move `CreateLinearMapper()` to the `disks` package
